+++ bin/actions/create_resource_request_action.js
@@ -2,20 +2,22 @@ const ReferenceParser = require(`../parsers/reference_parser`)
 const PackageFacade = require(`../facades/package_facade`)
 const TransitFacade = require(`../facades/transit_facade`)
 const DependencyLog = require('../support/dependency_log')
+const StatusLog = require('../support/status_log')
+const _ = require('lodash')
 
-module.exports = (reference) => {
+const createResourceRequestAction = (reference, installPath = undefined) => {
   return new Promise((resolve, reject) => {
-    const archiveRequest = ReferenceParser.referenceToArchiveRequest(reference)
-    const isRedundant = DependencyLog.hasRequest(archiveRequest)
+    const archiveRequest = ReferenceParser.referenceToArchiveRequest(reference, installPath)
+    const isRedundant = DependencyLog.hasInstalled(archiveRequest)
     const PackageTool = PackageFacade.of(archiveRequest)
     const TransitTool = TransitFacade.of(archiveRequest)
 
-    DependencyLog.trackInstallation(archiveRequest)
+    StatusLog.notify(`resource ${archiveRequest.uri}`, archiveRequest.uuid, _.merge({}, archiveRequest, {
+      packageTool: PackageTool.name,
+      transitTool: TransitTool.name
+    }))
 
-    // console.dir(_.merge(archiveRequest, {
-    //   package: PackageTool.name,
-    //   io: TransitTool.name
-    // }))
+    DependencyLog.trackInstallation(archiveRequest)
 
     resolve({
       archiveRequest,
@@ -25,3 +27,5 @@ module.exports = (reference) => {
     })
   })
 }
+
+module.exports = createResourceRequestAction