+++ bin/actions/download_archive_action.js
@@ -4,9 +4,9 @@ const DependencyLog = require('../support/dependency_log')
 const FileSystem = require('../support/file_system')
 const StatusLog = require('../support/status_log')
 
-const downloadArchiveAction = (reference) => {
+const downloadArchiveAction = (reference, installPath = undefined) => {
   // TODO cache with version name so we can pull multiple versions of one component to resolve requirements?
-  return createResourceRequestAction(reference)
+  return createResourceRequestAction(reference, installPath)
     .then((resourceRequest) => {
       const paths = applicationConfiguration.get(`paths`)
       const { isRedundant, archiveRequest, PackageTool, TransitTool } = resourceRequest
@@ -18,7 +18,6 @@ const downloadArchiveAction = (reference) => {
         })
       }
 
-      FileSystem.createDirectory(`${archiveRequest.stagingPath}/`)
       FileSystem.createDirectory(`${paths.cache}/`)
 
       StatusLog.notify(`cache ${archiveRequest.uri}`, archiveRequest.uuid)
@@ -27,6 +26,7 @@ const downloadArchiveAction = (reference) => {
           .then(({ availableVersions }) => {
             DependencyLog.trackAvailableVersions(archiveRequest, availableVersions)
             FileSystem.removeDirectory(`${archiveRequest.stagingPath}`)
+            FileSystem.createDirectory(`${archiveRequest.stagingPath}`)
 
             StatusLog.notify(`stage ${archiveRequest.uri}`, archiveRequest.uuid)
             return PackageTool