+++ bin/actions/install_archive_action.js
@@ -1,29 +1,32 @@
-const promoteArchiveAction = require('./promote_archive_action')
+const applicationConfiguration = require('../configurations/application')
+const ManifestConfiguration = require('../configurations/manifest')
+
+const FileSystem = require('../support/file_system')
 const StatusLog = require('../support/status_log')
+const _ = require('lodash')
 
-const installArchiveAction = (archiveRequest) => {
-  StatusLog.notify(`install ${archiveRequest.uri}`, archiveRequest.uuid)
-  return promoteArchiveAction(archiveRequest)
+const installArchiveAction = (archiveRequest, installedName) => {
+  const paths = applicationConfiguration.get(`paths`)
+  // TODO cache this?
+  const manifestJson = ManifestConfiguration.build(archiveRequest.stagingPath)
+  const archiveFolder = _.get(archiveRequest, 'scope.installation.destination', paths.archive)
+  // TODO not sure about this resolution - maybe just the installedName
+  const archivePath = `${archiveFolder}/${installedName || manifestJson.name || archiveRequest.archive}/`
+  const installFrom = `${archiveRequest.stagingPath}${_.get(archiveRequest, 'scope.installation.subfolder', ``)}`
+  const ignoreFolders = _.merge(
+    applicationConfiguration.get(`rules.ignoreFiles`, []),
+    manifestJson.ignore
+  )
 
-  // return downloadArchiveAction(reference)
-  //   .then(({ archiveRequest }) => {
-  //     StatusLog.notify(`install ${archiveRequest.uri}`, archiveRequest.uuid)
-  //     return promoteArchiveAction(archiveRequest)
-  //   })
+  StatusLog.notify(`install ${archivePath}`, archiveRequest.uuid)
+  return FileSystem
+    .createDirectory(`${archiveFolder}/`)
+    .removeDirectory(`${archiveFolder}/${archiveRequest.archive}`)
+    .copyNonIgnoredFiles(installFrom, archivePath, ignoreFolders)
+    .then(() => {
+      StatusLog.notify(`installed ${archivePath}`, archiveRequest.uuid)
+      return manifestJson
+    })
 }
 
 module.exports = installArchiveAction
-
-// StatusLog.notify(`version ${archiveRequest.uri}`, archiveRequest.uuid)
-// return VaultFacade
-//   .of(archiveRequest)
-//   .assignAppropriateVersion(archiveRequest)
-//     .then(() => {
-
-//     })
-
-// download - keep downloading
-// as you download for each dep request - add requirement to requirement tables
-// when done downloading, review conflicts
-// resolve conflicts
-// install each at one version -