+++ bin/parsers/reference_parser.js
@@ -1,3 +1,4 @@
+const crypto = require('crypto')
 const path = require('path')
 const _ = require('lodash')
 
@@ -5,7 +6,7 @@ const applicationConfiguration = require('../configurations/application')
 
 const Discover = {
   COMPONENT_ASPECT: /.*?\/([a-z0-9-_.]*?)[./]?(zip|tar|tgz|gz|tar.gz|git)?$/i,
-  IS_VERSION: /^[~^>=<]?\d{1,2}\.\d{1,2}\.\d{1,2}$/i
+  HAS_EMBEDDED_VERSION: /[]*?[_-]+([\d.]*)$/i
 }
 
 /**
@@ -18,10 +19,13 @@ const Discover = {
  * "uri#version" ===> { url, version }
  */
 class ReferenceParser {
-  static referenceToArchiveRequest (reference) {
+  static referenceToArchiveRequest(reference, overrideUniqueName = undefined) {
     const scopeOrReference = this.normalizeReference(reference)
-    const resource = this.scopeToResource(scopeOrReference)
-    const archiveRequest = this.resourceToArchiveRequest(resource)
+    const {
+      resource,
+      scope
+    } = this.__scopeToResource(scopeOrReference)
+    const archiveRequest = this.__resourceToArchiveRequest(resource, scope, overrideUniqueName)
 
     return archiveRequest
   }
@@ -33,100 +37,113 @@ class ReferenceParser {
    * folder res        => folder#master res
    *
    * @param { reference } reference
-   * @return "reference"
+   * @returns "reference"
    */
-  static normalizeReference (reference) {
+  static normalizeReference(reference) {
     return _.trimEnd(_.trimStart(reference || '', path.sep))
   }
 
-  /**
-   *
-   * @param {*} scopeOrReference
-   */
-  static scopeToResource (scopeOrReference) {
-    const patternMarkers = applicationConfiguration.get(`rules.patternMarkers`)
-    const [uri, version = ``] = this.splitURIVersion(scopeOrReference)
-    const uriAspects = uri.split(patternMarkers.separator)
-    const sourceConversionRule = this.__matchConversionRule(uriAspects)
-
-    if (!sourceConversionRule) {
-      return scopeOrReference
-    }
-
-    const {
-      pattern,
-      template,
-      constants
-    } = sourceConversionRule
-    const templateVariables = _.zipObject(pattern.split(patternMarkers.separator), uriAspects)
-
-    return _.template(template)(_.merge({}, templateVariables, constants, {
-      version
-    }))
+  static splitArchiveExtension(uri) {
+    const [_0, archive, extension] = Discover.COMPONENT_ASPECT.exec(uri) || [``, uri, ``] // eslint-disable-line
+    return [archive, extension]
   }
 
-  /**
-   *
-   * @param {*} reference
-   */
-  static resourceToArchiveRequest (reference) {
-    const versionMarker = _.first(applicationConfiguration.get(`rules.patternMarkers.version`))
-    const { staging, cache } = applicationConfiguration.get(`paths`)
-
-    const [uri, version = `master`] = this.splitURIVersion(reference)
-    const [archive, extension] = this.splitArchiveExtension(uri)
-
-    const safeExtension = (extension) ? `.${extension}` : `/`
-    const cachePath = `${cache}/${archive}__${version}${safeExtension}`
-    const stagingPath = `${staging}/${archive}/${version}/`
-    const uuid = `${archive}${versionMarker}${version}`
-
-    // TODO cachePath and stagingPath can avoid issues with folders named >=
-    // var crypto = require('crypto');
-    // crypto.createHash('md5').update(data).digest("hex");
-
-    return {
-      installedVersion: version,
-      stagingPath,
-      cachePath,
-      archive,
-      version,
-      uuid,
-      uri
-    }
-  }
-
-  /**
-   *
-   * @param {*} archiveRequest
-   * @param {*} archiveManifest
-   */
-  static buildArchivePath (archiveRequest, archiveManifest) {
-    const paths = applicationConfiguration.get(`paths`)
-    const archiveName = (archiveManifest.name) ? archiveManifest.name : archiveRequest.archive
-    const archiveFolder = archiveManifest.repositoryFolder || paths.archives
-
-    return `${archiveFolder}/${archiveName}/`
-  }
-
-  static splitURIVersion (reference) {
+  static splitURIVersion(reference) {
     const patternMarkers = applicationConfiguration.get(`rules.patternMarkers`)
     const versionMarker = _.find(patternMarkers.version, (versionMarker) => reference.indexOf(versionMarker) !== -1)
 
     return reference.split(versionMarker || _.first(patternMarkers.version))
   }
 
-  static splitArchiveExtension (uri) {
-    const [_0, archive, extension] = Discover.COMPONENT_ASPECT.exec(uri) || [``, uri, ``] // eslint-disable-line
-    return [archive, extension]
+  /**
+   *
+   * @param {*} scopeOrReference
+   */
+  static __scopeToResource(scopeOrReference) {
+    const patternMarkers = applicationConfiguration.get(`rules.patternMarkers`)
+    const [uri, version = ``] = this.splitURIVersion(scopeOrReference)
+    const uriAspects = uri.split(patternMarkers.separator)
+    const scope = this.__matchConversionRule(uriAspects)
+
+    if (!scope) {
+      return {
+        resource: scopeOrReference,
+        scope: {}
+      }
+    }
+
+    // TODO when we do `publish` we should use `push_uri` and configure this based on the operation
+    const {
+      pattern,
+      pull_uri,
+      constants
+    } = scope
+    const templateVariables = _.zipObject(pattern.split(patternMarkers.separator), uriAspects)
+
+    return {
+      scope,
+      resource: _.template(pull_uri)(_.merge({}, templateVariables, constants, {
+        version
+      }))
+    }
   }
 
   /**
    *
-   * @param {*} uri
-   * @param {*} uriAspects
+   * @param {*} resource
    */
-  static __matchConversionRule (uriAspects) {
+  static __resourceToArchiveRequest(resource, scope, overrideUniqueName = undefined) {
+    const versionMarker = _.first(applicationConfiguration.get(`rules.patternMarkers.version`))
+    const {
+      staging,
+      cache
+    } = applicationConfiguration.get(`paths`)
+
+    const [uri, _version] = this.splitURIVersion(resource)
+    const [_archive, extension] = this.splitArchiveExtension(uri)
+    // TODO should I default to * because master is for git?
+    const [archive, version = `master`] = this.__detectVersionInArchive(_archive, _version)
+
+    const versionFolder = crypto.createHash(`md5`).update(version).digest(`hex`)
+    const installedName = overrideUniqueName || archive
+
+    const safeExtension = (extension) ? `.${extension}` : `/`
+    const cachePath = `${cache}/${archive}__${versionFolder}${safeExtension}`
+    const stagingPath = `${staging}/${archive}/${versionFolder}/`
+    const uuid = `${installedName}${versionMarker}${version}`
+
+    // TODO pass unique name through to collect -v2 and v-3 and apply unique name
+
+    return {
+      installedVersion: version,
+      installedName,
+      stagingPath,
+      cachePath,
+      archive,
+      version,
+      scope,
+      uuid,
+      uri
+    }
+  }
+
+  static __detectVersionInArchive(archive, version) {
+    const embeddedVersion = Discover.HAS_EMBEDDED_VERSION.exec(archive)
+
+    if (embeddedVersion) {
+      return [archive.substr(0, embeddedVersion.index), _.last(embeddedVersion)]
+    } else {
+      return [archive, version]
+    }
+  }
+
+
+  /**
+   *
+   * @param {*} uriAspects
+   * @returns a scope object or undefined
+   */
+  static __matchConversionRule(uriAspects) {
     const sources = applicationConfiguration.get(`sources`)
     const scope = uriAspects[0] = (_.first(uriAspects) || '')