+++ bin/support/dependency_log.js
@@ -1,13 +1,10 @@
 const ReferenceParser = require('../parsers/reference_parser')
 const VersionServant = require('../servants/version_servant')
-const semver = require('semver')
 const _ = require('lodash')
 
 class DependencyLog {
   static initialize () {
-    // TODO build this list on start
     this.__installed = {}
-    this.__dependencies = {}
     this.__availableVersions = {}
     return this
   }
@@ -17,55 +14,59 @@ class DependencyLog {
     return this
   }
 
-  static trackAvailableVersions({ archive }, availableVersions) {
+  static trackAvailableVersions ({ archive }, availableVersions) {
     this.__availableVersions[archive] = availableVersions
     return this
   }
 
+  /**
+   *
+   * @param {*} archiveRequest uses installedName and version to identify a unique installation request
+   */
+  static hasInstalled ({ installedName, version }) {
+    return !!this.__installed[installedName] && !!this.__installed[installedName][version]
+  }
+
   static trackInstallation (archiveRequest) {
-    const { uuid } = archiveRequest
+    const { installedName, version } = archiveRequest
+    const installed = this.__installed
 
-    this.__installed[uuid] = this.__installed[uuid] || archiveRequest
+    installed[installedName] = installed[installedName] || {}
+    installed[installedName][version] = installed[installedName][version] || archiveRequest
     return this
-  }
-
-  static trackDependencies (uuid, dependencies) {
-    this.__dependencies[uuid] = dependencies
-    return this
-  }
-
-  static hasRequest ({ uuid }) {
-    return !!this.__installed[uuid]
   }
 
   static resolutions () {
-    const versionRequirements = DependencyLog.__calculateVersionRequirements()
+    const versionRequirements = this.__installed
     const versionMatches = DependencyLog.__calculateVersionMatches(versionRequirements)
     const versionConflicts = DependencyLog.__calculateVersionConflicts(versionRequirements)
 
-    const versionConflictsResolution = _.mapValues(versionConflicts, (conflicts, archive) => {
-      const availableVersions = this.__availableVersions[archive]
+    const versionConflictsResolution = _.mapValues(versionConflicts, (conflicts, installedName) => {
+      const archiveName = _.get(_.first(_.toArray(conflicts)), `archive`)
+      const availableVersions = this.__availableVersions[archiveName]
       const versionRange = _.keys(conflicts).join(' || ')
       const appropriateVersion = VersionServant.findAppropriateVersion(availableVersions, versionRange)
 
       if (!appropriateVersion) {
-        conflicts // what do I choose? it could:
+        // conflicts // what do I choose? it could:
         // >=3.2.3 <4.0.0 - if we do not match, then?
 
         // 1. get only versions from `conflicts` and choose which array is higher
         // 2. get the version (just version) that is highest
-        // 3. if it is a range - we... need the installed version so we can call findappriprivate just fro that rule and get a version...!
+        // 3. if it is a range - we... need the installed version so we can call find appropriate just fro that rule and get a version...!
+
+        const firstRequest = _.first(_.toArray(conflicts))
 
         // NOTE this is still an issue - what if it is a ~2.1.2
-        const mostRequested = _.findKey(conflicts, (requestedBy, version) => {
-          return requestedBy.length;
-        })
+        // const mostRequested = _.findKey(conflicts, (requestedBy, version) => {
+        //   return requestedBy.length
+        // })
 
         // TODO
-        const highestRequested = 0
+        // const highestRequested = 0
 
         // what if this is still null?!
-        return VersionServant.findAppropriateVersion(availableVersions, mostRequested)
+        return VersionServant.findAppropriateVersion(availableVersions, firstRequest)
       }
 
       return appropriateVersion
@@ -74,15 +75,19 @@ class DependencyLog {
     // TODO if we have this version - great. if not, then we need to get a new version - which means download again...
     // then we can copy folders
 
-    return _.merge({}, versionMatches, versionConflictsResolution)
+    return _.mapValues(_.merge({}, versionMatches, versionConflictsResolution), (version, installedName /* ISSUE might be archive */) => {
+      return _.get(this.__installed, `["${installedName}"]["${version}"]`, {
+        'warning': 'no version?!'
+      })
+    })
   }
 
   static __calculateVersionMatches (versionRequirements) {
     const versionMatches = {}
 
-    _.each(versionRequirements, (archiveRequirements, archive) => {
+    _.each(versionRequirements, (archiveRequirements, installedName) => {
       if (_.size(archiveRequirements) === 1) {
-        versionMatches[archive] = _.first(_.keys(archiveRequirements))
+        versionMatches[installedName] = _.first(_.keys(archiveRequirements))
       }
     })
 
@@ -92,42 +97,15 @@ class DependencyLog {
   static __calculateVersionConflicts (versionRequirements) {
     const versionConflicts = {}
 
-    _.each(versionRequirements, (archiveRequirements, archive) => {
+    _.each(versionRequirements, (archiveRequirements, installedName) => {
       if (_.size(archiveRequirements) > 1) {
-        versionConflicts[archive] = archiveRequirements
+        versionConflicts[installedName] = archiveRequirements
       }
     })
 
     return versionConflicts
   }
 
-  /**
-   * reverse map all dependencies to a map of requests
-   */
-  static __calculateVersionRequirements () {
-    const dependencies = this.__dependencies
-    const versionRequirements = {}
-
-    _.each(dependencies, (requirements, requestingArchive) => {
-      const [uri, version] = ReferenceParser.splitURIVersion(requestingArchive)
-      const [archive] = ReferenceParser.splitArchiveExtension(uri)
-
-      _.each(requirements, (requirement) => {
-        const [requirementURI, requirementVersion] = ReferenceParser.splitURIVersion(requirement)
-        const [requirementArchive] = ReferenceParser.splitArchiveExtension(requirementURI)
-
-        const _0 = versionRequirements[requirementArchive] = versionRequirements[requirementArchive] || {}
-        const _1 = _0[requirementVersion] = _0[requirementVersion] || []
-
-        _1.push({
-          archive,
-          version
-        })
-      })
-    })
-
-    return versionRequirements
-  }
 }
 
 module.exports = DependencyLog.initialize()