+++ bin/support/status_log.js
@@ -1,7 +1,7 @@
 const spinners = require('cli-spinners')
 const readline = require('readline')
+const winston = require('winston')
 const color = require('cli-color')
-const _ = require('lodash')
 
 const Colors = {
   gray: color.xterm(8),
@@ -13,33 +13,32 @@ const Spinner = {
 }
 
 class StatusLog {
-
-  static __drawLine(content) {
+  static __drawLine (content) {
     if (this.__stream) {
-      this.__stream.write(null, { ctrl: true, name: 'u' });
+      this.__stream.write(null, { ctrl: true, name: 'u' })
       this.__stream.write(content)
     }
   }
 
-  static __drawContent() {
+  static __drawContent () {
     const spinner = Spinner.running[this.__frame % Spinner.running.length]
     const seconds = parseInt(this.__frame / (1000 / this.__refreshRate), 10)
 
     return `${Colors.yellow(spinner)} ${Colors.gray(seconds + 's')} (${this.__action})`
   }
 
-  static start() {
+  static start () {
     this.__interval = setInterval(() => {
-      this.__frame += 1;
+      this.__frame += 1
       this.__drawLine(this.__drawContent())
     }, this.__refreshRate)
 
     return this
   }
 
-  static stop() {
+  static stop () {
     clearInterval(this.__interval)
-    if (!!this.__stream) {
+    if (this.__stream) {
       this.__stream.write(null, { ctrl: true, name: 'u' })
     }
 
@@ -48,6 +47,23 @@ class StatusLog {
 
   static initialize () {
     this.uninitialize()
+
+    this.__logger = new (winston.Logger)({
+      level: 'info',
+      exitOnError: false,
+      transports: [
+        new (winston.transports.File)({
+          timestamp: () => Date.now(),
+          handleExceptions: true,
+          filename: `vault.log`,
+          json: false,
+          options: {
+            flags: 'w'
+          }
+        })
+      ]
+    })
+
     this.__stream = readline.createInterface({
       output: process.stdout,
       input: process.stdin,
@@ -58,19 +74,23 @@ class StatusLog {
   }
 
   static uninitialize () {
+    this.__logger = undefined
     this.__stream = undefined
     this.__refreshRate = 250
     this.__frame = 0
-    this.__resources = {}
     this.__action = ''
     return this
   }
 
-  static notify (action, resource) {
-    if (!this.__resources[resource]) {
-      this.__resources[resource] = resource
-      this.__action = action
-    }
+  static notify (action, resource, meta = {}) {
+    this.__logger.info(`[${resource}] ${action}`, meta)
+    this.__action = action
+    return this
+  }
+
+  static error (action, resource, meta = {}) {
+    this.__logger.error(`[${resource}] ${action}`, meta)
+    this.__action = action
     return this
   }