+++ changelog.md
@@ -1,5 +1,13 @@
 # Klepto Changelog
 
+## 0.5.0
+
+- ADDED `install` with name resolution
+- ADDED `install` with installation folder and subfolder resolution
+- FIXED documentation from `@return` to `@returns`
+- REFACTORED the `uri` is parsed to see if the version is embedded in the archive name
+- ADDED logging to `vault.log` until the console messaging is clearer
+
 ## 0.4.0
 
 - ADDED support for version resolution